split conversion logic.
